cast changed

+++ include/Class_Local_Tree_2D.tpp
@@ -1830,7 +1830,7 @@ public:
 		if (oct->info[iface1] == false && oct->info[iface2] == false){
 
 			//Build Morton number of virtual neigh of same size
-			Class_Octant<2> samesizeoct(oct->level, oct->x+cx*size, oct->y+cy*size);
+			Class_Octant<2> samesizeoct(oct->level, int32_t(oct->x)+int32_t(cx)*int32_t(size), int32_t(oct->y)+int32_t(cy)*int32_t(size));
 			Morton = samesizeoct.computeMorton();
 
 			//SEARCH IN GHOSTS
